feat: database refactor (#33)
Update the database schema to normalize the data. Extract out the author, publisher, and vendor, and move the binding and genre to enums. Also switch the isbn to reference what it stores -- the ISBN 13 value (as a number to assure uniqueness).

This update was quite large, and we could have done this with migrations to the original schema. But because there was really no data of value, and we can drop the tables, starting fresh seemed simpler. So remove the old migrations and create just the single init.

The book action was updated quite a bit to handle these changes. There's a lot of TODOs in the code as there was just too much to take on in this first pass.

Add a seed script with an initial book and author. This can probably be improved with Faker for the values. But it's a start.

Finally, update the readme to include notes on this and other work.