feat: improve Book types (#36)
Remove the original Book type and replace it with either the BookCreateInput type (used to send input to create a book) or the BookHydrated type (response book value). This will set us up better for future work, and better standardizes on the types from Prisma. But it does have the downside of affecting a lot of files... So in addition to that change:

- The seeds, stories, UI components, pages, and actions have been updated to use the new types
- Add book page now internalizes the type which it exposes in the form.
- useExternalBookSearch provides "hints" as to what some of these values will be (since we may not match those values 1-1).
- BigInt serialization problems: In jest tests we need to apply "workerThreads" to fix the serialization issue. And in storybook, we have to convert to string for now. Comments added for when the issues are resolved.