chore: parse BigInt for serialization in nextjs (#39)
When the API responds with a value that includes a BigInt, it needs to parse it to JSON. This fails because JSON.stringify does not yet support BigInt. So provide an override as instructed by MDN. Include this in the next config so that it will be applied within the next framework (which includes the APIs).