feat: Add invoices to db schema (#56)
Add the concept of an Invoice to the database schema. Invoices represent receiving a shipment of books from a Vendor. The invoice will contain a list of InvoiceItems, each being a 1-1 with an actual Book. So by processing the Invoice, we'll add inventory to our Books.

This first pass just updates the database and resolves any issues with the code post-update. We'll later come in an add actions to create invoices and eventually update the UI with the new flow.