fix: fixes for invoice flow (#90)
There was a bug in the invoice completion flow, where we were writing updates for the same book independently within the transaction. So if the same book was in the list more than once, they would all use the same original value for quantity. This fix combines the same books to a single update to fix this issue.

Additional fixes found during testing:
- since we're not using pagination in the UI, increase the limit to 100 until we do
- correctly show the discount percentage string, adding a function in money