feat: ProductType (#106)
In the future, we (may) want to sell things other than books. So to build for that now, add a ProductType enum to the database schema with a single value BOOK. Then within InvoiceItem, include the ProductType and make the book/bookId optional.

Make updates to the code to account for this new ability, mostly skipping types other than BOOK since we have no other type at this time.

Adjust the InvoiceItemsTable to be more generic in how it describes each item, add an image to the table, and add stories to demo.