feat: Format/Genre enum -> model
Move Format and Genre from enums to models in the prisma schema, thus making a Format and Genre table. This will allow for more flexibility in how we describe the book formats and genres, and perform basic CRUD operations to them.

This change is a large one! Originally done separately: one for format and one for genre. But the changes were so similar that it made more sense to do it all at once. Tests were added where they were lacking, and the e2e flow was run manually to verify (need better playwright tests).

In addition, the seeds were updated as part of this change, since we now have "core" seed data for the basic formats and genres. The "fake" seed data generates the core data prior to generating the fakes.