feat: checkout page (#132)
Build out the rough idea for the checkout page. This page starts with an input field to enter in a SKU. When submitted, will search for the book, create the order (if not yet created), create the order item, and re-load the order. Subsequent searches will just create additional order items.