feat: search handles ISBN number input
We're expanding the book search capability to support a number as input (in the string), and if so, assume it's an ISBN13 value. We're at the point where we may want to consider making this search input a bit more specific to what it's looking for (perhaps an object rather than string?), but it works for now.

If the input is a number, query for a matching `isbn13` value. Add a test to demo.