feat: process transaction page
Add page which is responsible for moving an order from OPEN to PAID. This will most likely involve creating a transaction in the background, and waiting for the transaction to complete (the customer to make the purchase). While we wait, show a loading animation to the admin and provide a way to cancel the transaction if needed.

For now, this is mostly just for show.