test: checkout e2e test
Add an end to end test for checkout. This requires some inventory to begin with, so update our CI seed script to add books.

The test adds that book to the order, then pays with square, and verifies the order entry exists along with subtracting the inventory from the book.

For square, this means we must be setup in sandbox for this test, with a setup that will mark the transaction complete for orders totaling less than $25.